+++ src/dv-object/infra/repositories/transformers/OwnerPayload.ts
@@ -1,0 +1,6 @@
+export interface OwnerPayload {
+  type: string;
+  identifier: string;
+  displayName: string;
+  owner?: OwnerPayload;
+}
